requirements.txt
Django~=4.2
djangorestframework~=3.14.0
django-ckeditor~=6.5.1
Pillow~=9.5.0
pandas~=2.0.1
sklearn~=0.0.post4
scikit-learn~=1.2.2
tensorflow~=2.12.0
numpy~=1.24.3
dj-rest-auth~=3.0.0
django-allauth~=0.54.0